move goal logic from application glue logic
